Extract OTLP trace test helpers
Move trace-specific fixtures into OtlpTraceUtils so shared OTLP test helpers stay focused and the synthetic trace/span IDs are named clearly for reviewers.
